+++ test/unified_format.py
@@ -72,13 +72,9 @@ from pymongo.errors import (
     ConfigurationError,
     ConnectionFailure,
     EncryptionError,
-    ExecutionTimeout,
     InvalidOperation,
-    NetworkTimeout,
     NotPrimaryError,
     PyMongoError,
-    ServerSelectionTimeoutError,
-    WriteConcernError,
 )
 from pymongo.monitoring import (
     _SENSITIVE_COMMANDS,
@@ -952,13 +948,8 @@ class UnifiedSpecTestMixinV1(IntegrationTest):
                 self.assertNotIsInstance(exception, PyMongoError)
 
         if is_timeout_error:
-            # TODO: PYTHON-3291 Implement error transformation.
-            if isinstance(exception, WriteConcernError):
-                self.assertEqual(exception.code, 50)
-            else:
-                self.assertIsInstance(
-                    exception, (NetworkTimeout, ExecutionTimeout, ServerSelectionTimeoutError)
-                )
+            self.assertIsInstance(exception, PyMongoError)
+            self.assertTrue(exception.timeout, msg=exception)
 
         if error_contains:
             if isinstance(exception, BulkWriteError):